+++ atdts/test/ts-tests/test_atdts.ts
@@ -54,7 +54,8 @@ function test_everything() {
       ["h", 8],
     ]),
     options: [{value:10}, null, {value:88}],
-    nullables: [13, 71]
+    nullables: [13, 71],
+    untyped_things: [{}, [["hello"]], 123]
   }
   const a_str = JSON.stringify(API.writeRoot(a_obj), null, 2)
   save('a_str', a_str)
@@ -147,6 +148,15 @@ ${a_str}`
   "nullables": [
     13,
     71
+  ],
+  "untyped_things": [
+    {},
+    [
+      [
+        "hello"
+      ]
+    ],
+    123
   ]
 }`
   save('b_str', b_str)